startup scripts

+++ startup_with_snapshot.sh
@@ -6,7 +6,8 @@ if [[ $? == 0 ]]; then
     exit 0
 fi
 
-touch startup.temp
+## source will load the variable written later in the script if it exists
+source startup.temp
 
 if [[ $docker_up != 1 ]]; then
     echo "Setting Up Startup..."
@@ -46,6 +47,13 @@ if [[ $docker_up != 1 ]]; then
             modified_HAF_SHM="HAF_SHM_DIRECTORY=\"/mnt/haf_shared_mem\""
 
         fi
+        if [[ $line == ARGUMENTS=* && ($line == *--replay-blockchain* || $line == *--force-replay*) ]]; then
+            original_arguments="$line"
+            modified_arguments=$(echo "$line" | sed -E 's/(--replay-blockchain|--force-replay)//g')
+            echo "Using: $original_arguments"
+            echo "After Sync will use: $modified_arguments"
+            echo "If this isn't desired, manually change the arguments in .env before the sync is finished"
+        fi
     done < .env
 
     sed -i "s/$original_line/$modified_line/g" .env
@@ -58,6 +66,11 @@ if [[ $docker_up != 1 ]]; then
 
     echo "original_line=$original_line" >> startup.temp
     echo "modified_line=$modified_line" >> startup.temp
+
+    if [[ $original_arguments != "" ]]; then
+        echo "original_arguments=$original_arguments" >> startup.temp
+        echo "modified_arguments=$modified_arguments" >> startup.temp
+    fi
 
     docker compose up -d
 
@@ -73,29 +86,46 @@ docker compose logs -f | while read -r line; do
         echo "$line"
     fi
     if [[ $line == *"PROFILE: Entered LIVE sync"* ]]; then
+        sync_done=1
         echo "Detected *PROFILE: Entered LIVE sync* in the output. Bringing down Docker Compose..."
+
+        # Write Sync time to "haf.log" for tracking, as this log will get wiped on restart
         docker logs haf-world-haf-1 | grep PRO > haf.log
         docker compose down
         break
     fi
 done
 
-# Restore the original line
-sed -i "s/$modified_line/$original_line/g" .env
 
-# Move the shared_mem file to the blockchain directory
-if [[ $remove_shared_mem == 1 ]]; then
-    sed -i "s#$modified_HAF_SHM#$original_HAF_SHM#g" .env
-    sudo cp /mnt/haf_shared_mem/shared_memory.bin /haf-pool/haf-datadir/blockchain
-    sudo chown 1000:100 /haf-pool/haf-datadir/blockchain/shared_memory.bin
-    sudo umount /mnt/haf_shared_mem
+# prevent script completion if interupt sent to above loop
+
+if [[ $sync_done == 1 ]]; then
+    # Restore the original line
+    sed -i "s/$modified_line/$original_line/g" .env
+
+    # Move the shared_mem file to the blockchain directory
+    if [[ $remove_shared_mem == 1 ]]; then
+        sed -i "s#$modified_HAF_SHM#$original_HAF_SHM#g" .env
+        sudo cp /mnt/haf_shared_mem/shared_memory.bin /haf-pool/haf-datadir/blockchain
+        sudo chown 1000:100 /haf-pool/haf-datadir/blockchain/shared_memory.bin
+        sudo umount /mnt/haf_shared_mem
+    fi
+
+
+    # Remove replay arguments
+    if [[ $original_arguments != "" ]]; then
+        sed -i "s#$original_arguments#$modified_arguments#g" .env
+    fi
+
+    # Create a snapshot of the ZFS pool
+    ./snapshot_zfs_datasets.sh first_sync
+
+    # Restart Docker Compose
+    docker compose up -d
+    rm startup.temp
+
+    exit 0
 fi
 
-# Create a snapshot of the ZFS pool
-./snapshot_zfs_datasets.sh first_sync
+exit 1
 
-# Restart Docker Compose
-docker compose up -d
-rm startup.temp
-
-exit 0
